requirements.txt
# Create/update requirements.txt
cat > requirements.txt << EOF
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Payments
stripe==7.4.0

# Screenshot Engine (THIS IS THE KEY ONE!)
playwright==1.40.0

# Storage
boto3==1.34.10

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
orjson==3.9.10
itsdangerous==2.1.2
EOF

# Install all
pip install -r requirements.txt

# Install Playwright browsers
playwright install chromium